[Silabs] Rename Wifi interface sources files (#36318)
* Rename Wifi interface sources files

* restyle
