+++ examples/platform/silabs/efr32/rs911x/hal/efx_spi.c
@@ -41,10 +41,10 @@
 #include "sl_status.h"
 #include "spidrv.h"
 
+#include "WifiInterfaceAbstraction.h"
 #include "silabs_utils.h"
 #include "spi_multiplex.h"
 #include "wfx_host_events.h"
-#include "wfx_rsi.h"
 
 #ifdef SL_BOARD_NAME
 #include "sl_board_control.h"
