+++ examples/platform/silabs/efr32/rs911x/hal/rsi_hal_mcu_ioports.c
@@ -35,8 +35,8 @@
 #include "event_groups.h"
 #include "task.h"
 
+#include "WifiInterfaceAbstraction.h"
 #include "wfx_host_events.h"
-#include "wfx_rsi.h"
 
 #include "rsi_board_configuration.h"
 #include "rsi_driver.h"
