+++ examples/platform/silabs/efr32/rs911x/hal/rsi_hal_mcu_timer.c
@@ -39,6 +39,7 @@ extern void xPortSysTickHandler(void);
 /* RSI Driver include file */
 #include "rsi_driver.h"
 /* RSI WLAN Config include file */
+#include "WifiInterfaceAbstraction.h"
 #include "rsi_bootup_config.h"
 #include "rsi_common_apis.h"
 #include "rsi_data_types.h"
@@ -49,7 +50,6 @@ extern void xPortSysTickHandler(void);
 #include "rsi_wlan.h"
 #include "rsi_wlan_apis.h"
 #include "rsi_wlan_config.h"
-#include "wfx_rsi.h"
 
 #ifndef _use_the_rsi_defined_functions
