+++ examples/platform/silabs/efr32/rs911x/hal/sl_si91x_ncp_utility.c
@@ -44,10 +44,10 @@
 #include "sl_device_init_hfxo.h"
 #include "sl_status.h"
 
+#include "WifiInterfaceAbstraction.h"
 #include "silabs_utils.h"
 #include "sl_si91x_ncp_utility.h"
 #include "wfx_host_events.h"
-#include "wfx_rsi.h"
 
 #if SL_MX25CTRL_MUX
 sl_status_t sl_wfx_host_spiflash_cs_assert(void);
