+++ examples/platform/silabs/efr32/rs911x/rs911x.gni
@@ -1,15 +1,28 @@
+# Copyright (c) 2024 Project CHIP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import("//build_overrides/chip.gni")
 import("//build_overrides/efr32_sdk.gni")
 import("${silabs_sdk_build_root}/efr32_sdk.gni")
 
 rs911x_src_plat = [
-  "${examples_plat_dir}/rs911x/rsi_if.cpp",
+  "${examples_plat_dir}/rs911x/Rsi91xWifiInterface.cpp",
   "${examples_plat_dir}/rs911x/hal/rsi_hal_mcu_interrupt.c",
   "${examples_plat_dir}/rs911x/hal/rsi_hal_mcu_ioports.c",
   "${examples_plat_dir}/rs911x/hal/rsi_hal_mcu_timer.c",
   "${examples_plat_dir}/rs911x/hal/efx_spi.c",
-  "${silabs_common_plat_dir}/wifi/wfx_notify.cpp",
-  "${silabs_common_plat_dir}/wifi/wfx_rsi_host.cpp",
+  "${silabs_common_plat_dir}/wifi/WifiInterfaceAbstraction.cpp",
 ]
 
 #
@@ -59,4 +72,5 @@ rs911x_inc_plat = [
   "${examples_plat_dir}/rs911x",
   "${examples_plat_dir}/rs911x/hal",
   "${wiseconnect_sdk_root}/sapi/include",
+  "${silabs_common_plat_dir}/wifi",
 ]
