+++ examples/platform/silabs/efr32/wf200/wf200.gni
@@ -1,12 +1,30 @@
+# Copyright (c) 2024 Project CHIP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import("//build_overrides/chip.gni")
 import("//build_overrides/efr32_sdk.gni")
 import("${silabs_sdk_build_root}/efr32_sdk.gni")
 
-wf200_plat_incs = [ "${examples_plat_dir}/wf200" ]
+wf200_plat_incs = [
+  "${examples_plat_dir}/wf200",
+  "${silabs_common_plat_dir}/wifi",
+]
+
 wf200_plat_src = [
-  "${silabs_common_plat_dir}/wifi/wfx_notify.cpp",
   "${examples_plat_dir}/wf200/sl_wfx_task.c",
   "${examples_plat_dir}/wf200/wf200_init.c",
   "${examples_plat_dir}/wf200/efr_spi.c",
-  "${examples_plat_dir}/wf200/host_if.cpp",
+  "${examples_plat_dir}/wf200/Wf200WifiInterface.cpp",
+  "${silabs_common_plat_dir}/wifi/WifiInterfaceAbstraction.cpp",
 ]
