+++ src/platform/silabs/rs911x/wfx_sl_ble_init.h
@@ -22,14 +22,10 @@
  */
 #pragma once
 
-/**
- * Include files
- * */
 // BLE include file to refer BLE APIs
 #include "ble_config.h"
 #include "cmsis_os2.h"
 #include "wfx_host_events.h"
-#include "wfx_rsi.h"
 #include <rsi_ble.h>
 #include <rsi_ble_apis.h>
 #include <rsi_ble_common_config.h>
